refactor(Pet): change weight BigDecimal -> float - Pet weight doesn't need precise calculation - Updated DTO and database column type to float

+++ src/main/java/com/kakao/termproject/pet/dto/PetUpdateRequest.java
@@ -1,7 +1,6 @@
 package com.kakao.termproject.pet.dto;
 
 import com.kakao.termproject.pet.domain.Gender;
-import java.math.BigDecimal;
 
 public record PetUpdateRequest(
   String name,
@@ -10,7 +9,7 @@ public record PetUpdateRequest(
   String birthDate,
   boolean neutralize,
   boolean vaccinated,
-  BigDecimal weight,
+  float weight,
   String preferredWeather,
   String chronicDisease,
   String preferredPath
